+++ src/call_python.cpp
@@ -12,12 +12,13 @@ namespace py = pybind11;
 class MexFunction : public matlab::mex::Function {
 
     protected:
-        py::tuple convMat2np(ArgumentList inputs, py::handle owner, size_t lastInd, libpymcr::pymat_converter* converter) {
+        libpymcr::pymat_converter *_converter = nullptr;
+        py::tuple convMat2np(ArgumentList inputs, py::handle owner, size_t lastInd) {
             // Note that this function must be called when we have the GIL
             size_t narg = inputs.size() + lastInd;
             py::tuple retval(narg);
             for (size_t idx = 1; idx <= narg; idx++) {
-                retval[idx - 1] = converter->to_python(inputs[idx]);
+                retval[idx - 1] = _converter->to_python(inputs[idx]);
             }
             return retval;
         }
@@ -25,17 +26,20 @@ class MexFunction : public matlab::mex::Function {
     public:
         void operator()(ArgumentList outputs, ArgumentList inputs) {
             matlab::data::ArrayFactory factory;
-            if (inputs.size() < 1 || inputs[0].getNumberOfElements() != 2 ||
-                inputs[0].getType() != matlab::data::ArrayType::UINT64) { // Matlab only supports 64-bit
-                    throw std::runtime_error("First input must be pointers to a Python function and converter.");
+            if (inputs.size() < 1 || inputs[0].getType() != matlab::data::ArrayType::CHAR) {
+                throw std::runtime_error("Input must be reference to a Python function.");
             }
-            uintptr_t key = inputs[0][0];
-            uintptr_t conv_addr = inputs[0][1];
+            matlab::data::CharArray key = inputs[0];
 
             PyGILState_STATE gstate = PyGILState_Ensure();  // GIL{
 
-            libpymcr::pymat_converter* converter = reinterpret_cast<libpymcr::pymat_converter*>(conv_addr);
-            PyObject* fn_ptr = reinterpret_cast<PyObject*>(key);
+            if (_converter == nullptr) {
+                _converter = new libpymcr::pymat_converter(libpymcr::pymat_converter::NumpyConversion::WRAP);
+            }
+
+            py::module pyHoraceFn = py::module::import("libpymcr");
+            py::dict fnDict = pyHoraceFn.attr("_globalFunctionDict");
+            PyObject *fn_ptr = PyDict_GetItemString(fnDict.ptr(), key.toAscii().c_str());
 
             if (PyCallable_Check(fn_ptr)) {
                 PyObject *result;
@@ -44,12 +48,12 @@ class MexFunction : public matlab::mex::Function {
                     try {
                         const matlab::data::StructArray in_struct(inputs[endIdx]);
                         const matlab::data::Array v = in_struct[0][matlab::data::MATLABFieldIdentifier("pyHorace_pyKwArgs")];
-                        PyObject *kwargs = converter->to_python(inputs[endIdx]);
+                        PyObject *kwargs = _converter->to_python(inputs[endIdx]);
                         PyDict_DelItemString(kwargs, "pyHorace_pyKwArgs");
-                        py::tuple arr_in = convMat2np(inputs, fn_ptr, -2, converter);
+                        py::tuple arr_in = convMat2np(inputs, fn_ptr, -2);
                         result = PyObject_Call(fn_ptr, arr_in.ptr(), kwargs);
                     } catch (...) {
-                        py::tuple arr_in = convMat2np(inputs, fn_ptr, -1, converter);
+                        py::tuple arr_in = convMat2np(inputs, fn_ptr, -1);
                         result = PyObject_CallObject(fn_ptr, arr_in.ptr());
                     }
                 } catch (...) {
@@ -61,7 +65,7 @@ class MexFunction : public matlab::mex::Function {
                 }
                 else {
                     try {
-                        outputs[0] = converter->to_matlab(result, Py_REFCNT(result)==1);
+                        outputs[0] = _converter->to_matlab(result, Py_REFCNT(result)==1);
                     } catch (char *e) {
                         Py_DECREF(result);
                         PyGILState_Release(gstate);